https://ndphong.atlassian.net/browse/A2G-3
